fix(lut): match C float/double evaluation in atmcorlamb2 LUT chain
The scene-center atmospheric coefficients are cubic least-squares fits over
22 atmcorlamb2 evaluations. The fit is poorly conditioned, so 1-2 ULP
differences in roatm/ttatmg/satm became 100s-10000s of ULPs in the
coefficients and flipped aerosol retrieval ties (e.g. AOT 2.8 vs 3.0 at
high solar zenith), which the invalid-aerosol fill then propagated across
large areas.

Mirror C's per-expression types in comptg, local_chand, comproatm,
compsalb/comptrans and the scattering angle: float inputs, double math.h
calls and double literals where C promotes, and float storage. Notable:
- atmcorlamb2 uses a double lambda_sf (atmcorlamb2_new uses float)
- the log-AOT interpolation weight is computed in double
- isca/sca1 are computed in double
- compute_gas_transmission now takes and returns f32

Also fix the view zenith transmittance lookup, which used the solar zenith
grid (xtsmin/xtsstep) instead of C's view grid (xtvmin/xtvstep).

Validated against the C functions compiled standalone: all 7 bands x 22 AOT
values of roatm, ttatmg, satm and tgo are bit-identical for
LC09_L1TP_035024_20250104.

+++ lasrc_core/src/geometry.rs
@@ -91,10 +91,11 @@ pub fn scattering_angle(xmus: f64, xmuv: f64, cosxfi: f64) -> f64 {
     let xmus_f = xmus as f32;
     let xmuv_f = xmuv as f32;
     let cosxfi_f = cosxfi as f32;
-    let cscaa: f32 = (-(xmus_f as f64) * xmuv_f as f64
+    // -xmus * xmuv and xmus * xmus are float products; the rest is double.
+    let cscaa: f32 = ((-xmus_f * xmuv_f) as f64
         - cosxfi_f as f64
-            * (1.0 - xmus_f as f64 * xmus_f as f64).sqrt()
-            * (1.0 - xmuv_f as f64 * xmuv_f as f64).sqrt()) as f32;
+            * (1.0 - (xmus_f * xmus_f) as f64).sqrt()
+            * (1.0 - (xmuv_f * xmuv_f) as f64).sqrt()) as f32;
     let cscaa = cscaa.clamp(-1.0, 1.0);
     let scaa: f32 = ((cscaa as f64).acos() * RAD2DEG) as f32;
     scaa as f64
